refactor(tests): remove duplicate test files, rename consolidated
- Delete 7 original test files that were duplicates of consolidated versions:
  test_core_json.py, test_hashing.py, test_importers_chatgpt.py,
  test_importers_claude.py, test_projections.py, test_rendering_core.py,
  test_search.py

- Rename 10 *_consolidated.py files to drop meaningless suffix

- Removed 331 duplicate tests (1839 → 1507 total)
  Same coverage, no redundancy
