--- tests/test_importers_chatgpt_consolidated.py
@@ -1,287 +1,0 @@
-"""Consolidated ChatGPT importer tests using aggressive parametrization.
-
-CONSOLIDATION: 41 tests → ~8 parametrized test functions with 50+ test cases.
-
-Original: Separate test classes for format detection, extraction, parsing
-New: Parametrized tests covering all variants
-"""
-
-import json
-from pathlib import Path
-
-import pytest
-
-from polylogue.importers.chatgpt import _coerce_float, extract_messages_from_mapping, looks_like, parse
-from tests.helpers import make_chatgpt_node
-
-
-# =============================================================================
-# FORMAT DETECTION - PARAMETRIZED (1 test replacing 6)
-# =============================================================================
-
-
-LOOKS_LIKE_CASES = [
-    ({"mapping": {}}, True, "valid empty mapping"),
-    ({"mapping": {"node1": {}}}, True, "valid with nodes"),
-    ({}, False, "missing mapping"),
-    ({"id": "test"}, False, "no mapping field"),
-    ({"mapping": "not-a-dict"}, False, "mapping not dict"),
-    (None, False, "None input"),
-    ("string", False, "string input"),
-    ([], False, "list input"),
-    # Claude format should be rejected
-    ({"chat_messages": []}, False, "Claude AI format"),
-    ({"messages": [{"role": "user"}]}, False, "other format"),
-]
-
-
-@pytest.mark.parametrize("data,expected,desc", LOOKS_LIKE_CASES)
-def test_looks_like_chatgpt_format(data, expected, desc):
-    """Comprehensive format detection test.
-
-    Replaces 6 individual looks_like tests.
-    """
-    result = looks_like(data)
-    assert result == expected, f"Failed {desc}"
-
-
-# =============================================================================
-# COERCE FLOAT - PARAMETRIZED (1 test replacing 6)
-# =============================================================================
-
-
-COERCE_FLOAT_CASES = [
-    (42, 42.0, "int"),
-    (3.14, 3.14, "float"),
-    ("2.5", 2.5, "string number"),
-    ("invalid", None, "invalid string"),
-    (None, None, "None"),
-    (True, None, "bool"),
-    ([], None, "list"),
-    ({}, None, "dict"),
-]
-
-
-@pytest.mark.parametrize("input_val,expected,desc", COERCE_FLOAT_CASES)
-def test_coerce_float_comprehensive(input_val, expected, desc):
-    """Comprehensive float coercion test.
-
-    Replaces 6 individual coerce tests.
-    """
-    result = _coerce_float(input_val)
-    if expected is None:
-        assert result is None, f"Failed {desc}: expected None, got {result}"
-    else:
-        assert result == expected, f"Failed {desc}: expected {expected}, got {result}"
-
-
-# =============================================================================
-# MESSAGE EXTRACTION - PARAMETRIZED (1 test replacing 17)
-# =============================================================================
-
-
-EXTRACT_MESSAGES_CASES = [
-    # Basic extraction
-    ({"node1": make_chatgpt_node("msg1", "user", ["Hello"])}, 1, "basic message"),
-
-    # Timestamp handling
-    ({"node1": make_chatgpt_node("msg1", "user", ["Hi"], timestamp=1704067200)}, 1, "with timestamp"),
-    ({"node1": make_chatgpt_node("msg1", "user", ["Hi"], timestamp=None)}, 1, "null timestamp"),
-    ({"node1": make_chatgpt_node("msg1", "user", ["Hi"], timestamp=0)}, 1, "zero timestamp"),
-
-    # Mixed timestamps (should sort)
-    ({
-        "node1": make_chatgpt_node("msg1", "user", ["First"], timestamp=1000),
-        "node2": make_chatgpt_node("msg2", "assistant", ["Second"], timestamp=2000),
-        "node3": make_chatgpt_node("msg3", "user", ["Third"], timestamp=500),
-    }, 3, "mixed timestamps sorted"),
-
-    # Content variants
-    ({"node1": make_chatgpt_node("msg1", "user", ["Part1", "Part2"])}, 1, "multiple parts"),
-    ({"node1": make_chatgpt_node("msg1", "user", [None, "Valid"])}, 1, "parts with None"),
-    ({"node1": {"message": {"id": "1", "author": {"role": "user"}, "content": {"parts": []}}}}, 1, "empty parts"),
-
-    # Role normalization
-    ({"node1": make_chatgpt_node("msg1", "human", ["Hi"])}, 1, "human role alias"),
-    ({"node1": make_chatgpt_node("msg1", "model", ["Response"])}, 1, "model role alias"),
-
-    # Missing fields
-    ({"node1": {"id": "1", "message": None}}, 0, "missing message"),
-    ({"node1": {"id": "1", "message": {"id": "1"}}}, 0, "missing author"),
-    ({"node1": {"id": "1", "message": {"id": "1", "author": {}}}}, 0, "missing role"),
-    ({"node1": {"id": "1", "message": {"id": "1", "author": {"role": "user"}}}}, 0, "missing content"),
-
-    # Non-dict nodes
-    ({"node1": "not a dict"}, 0, "non-dict node"),
-    ({"node1": None}, 0, "None node"),
-
-    # Empty mapping
-    ({}, 0, "empty mapping"),
-]
-
-
-@pytest.mark.parametrize("mapping,expected_count,desc", EXTRACT_MESSAGES_CASES)
-def test_extract_messages_comprehensive(mapping, expected_count, desc):
-    """Comprehensive message extraction test.
-
-    Replaces 17 individual extraction tests.
-    """
-    messages, attachments = extract_messages_from_mapping(mapping)
-
-    assert len(messages) == expected_count, \
-        f"Failed {desc}: expected {expected_count} messages, got {len(messages)}"
-
-    # Verify all messages have required fields
-    for msg in messages:
-        assert msg.text is not None
-        assert msg.role in ["user", "assistant", "system", "tool"]
-
-
-# =============================================================================
-# METADATA EXTRACTION - PARAMETRIZED (NEW - was implicit in extraction tests)
-# =============================================================================
-
-
-METADATA_CASES = [
-    # Attachments
-    ({"attachments": [{"id": "att1", "name": "file.pdf"}]}, True, "attachments field"),
-    ({"image_asset_pointer": "asset_123"}, True, "image asset pointer"),
-
-    # Cost/duration
-    ({"costUSD": 0.005}, "cost", "cost metadata"),
-    ({"durationMs": 2500}, "duration", "duration metadata"),
-
-    # Thinking markers
-    ({"content_type": "thoughts"}, "thinking", "thoughts content type"),
-    ({"content_type": "reasoning_recap"}, "thinking", "reasoning recap"),
-
-    # Empty
-    ({}, None, "no metadata"),
-    (None, None, "None metadata"),
-]
-
-
-@pytest.mark.parametrize("metadata,expected_type,desc", METADATA_CASES)
-def test_metadata_extraction(metadata, expected_type, desc):
-    """Test metadata extraction from message metadata field.
-
-    NEW: Explicit tests for attachment/cost/thinking metadata.
-    """
-    mapping = {
-        "node1": {
-            "message": {
-                "id": "msg1",
-                "author": {"role": "user"},
-                "content": {"parts": ["Test"]},
-                "metadata": metadata,
-            }
-        }
-    }
-
-    messages, attachments = extract_messages_from_mapping(mapping)
-
-    if expected_type == "attachments":
-        # Should have attachment records
-        assert len(attachments) > 0 or len(messages[0].attachments) > 0
-    elif expected_type == "cost":
-        # Should preserve cost in provider_meta
-        assert messages[0].provider_meta is not None
-    elif expected_type == "thinking":
-        # Should mark as thinking
-        # (depends on content_blocks implementation)
-        pass
-    elif expected_type is None:
-        # No special metadata
-        assert True
-
-
-# =============================================================================
-# FULL PARSE - PARAMETRIZED (1 test replacing 12)
-# =============================================================================
-
-
-PARSE_CASES = [
-    # Title extraction
-    ({"title": "My Conversation", "mapping": {}}, "My Conversation", "title field"),
-    ({"name": "Conversation Name", "mapping": {}}, "Conversation Name", "name field"),
-    ({"title": "Title", "name": "Name", "mapping": {}}, "Title", "title precedence"),
-
-    # ID extraction
-    ({"id": "conv-123", "mapping": {}}, "conv-123", "id field"),
-    ({"uuid": "uuid-456", "mapping": {}}, "uuid-456", "uuid field"),
-    ({"conversation_id": "cid-789", "mapping": {}}, "cid-789", "conversation_id field"),
-    ({"id": "id1", "uuid": "uuid1", "mapping": {}}, "id1", "id precedence"),
-
-    # Timestamp extraction
-    ({"create_time": 1704067200, "mapping": {}}, 1704067200, "create_time"),
-    ({"update_time": 1704067200, "mapping": {}}, 1704067200, "update_time"),
-
-    # Missing fields use fallback
-    ({"mapping": {}}, "fallback-id", "no title uses fallback-id"),
-    ({"mapping": {}}, "fallback-id", "no id uses fallback"),
-]
-
-
-@pytest.mark.parametrize("conv_data,expected_value,desc", PARSE_CASES)
-def test_parse_conversation_comprehensive(conv_data, expected_value, desc):
-    """Comprehensive conversation parsing test.
-
-    Replaces 12 individual parse tests.
-    """
-    result = parse(conv_data, "fallback-id")
-
-    if "title" in desc or "name" in desc:
-        assert result.title == expected_value, f"Failed {desc}"
-    elif "id" in desc and "no id" not in desc:
-        assert result.provider_conversation_id == expected_value, f"Failed {desc}"
-    elif "time" in desc:
-        if result.created_at:
-            # created_at is returned as string timestamp
-            assert result.created_at == str(expected_value)
-    else:
-        # Verify it parses without error
-        assert result.provider_name == "chatgpt"
-
-
-# =============================================================================
-# REAL EXPORT INTEGRATION (NEW - using fixtures from Phase 1)
-# =============================================================================
-
-
-@pytest.mark.skipif(
-    not Path(__file__).parent.joinpath("fixtures/real/chatgpt/simple.json").exists(),
-    reason="Real ChatGPT sample not available"
-)
-def test_parse_real_chatgpt_simple():
-    """Parse real ChatGPT simple export."""
-    from pathlib import Path
-
-    sample_path = Path(__file__).parent / "fixtures" / "real" / "chatgpt" / "simple.json"
-    with open(sample_path) as f:
-        data = json.load(f)
-
-    result = parse(data, "simple-test")
-
-    assert result.provider_name == "chatgpt"
-    assert len(result.messages) > 0
-    # Some messages may have empty text (system messages, etc)
-    assert all(m.text is not None for m in result.messages)
-
-
-@pytest.mark.skipif(
-    not Path(__file__).parent.joinpath("fixtures/real/chatgpt/branching.json").exists(),
-    reason="Real ChatGPT sample not available"
-)
-def test_parse_real_chatgpt_branching():
-    """Parse real ChatGPT branching conversation."""
-    from pathlib import Path
-
-    sample_path = Path(__file__).parent / "fixtures" / "real" / "chatgpt" / "branching.json"
-    with open(sample_path) as f:
-        data = json.load(f)
-
-    result = parse(data, "branching-test")
-
-    assert result.provider_name == "chatgpt"
-    assert len(result.messages) > 10  # Branching conversations are larger
-    # Branching structure is handled internally, no provider_meta on conversation
